refactor: deprecate `nuxt serve` in favor of `nuxt start`
