+++ packages/cli/src/commands/export.js
@@ -4,7 +4,7 @@ import generate from './generate'
 export default {
   ...generate,
   run (...args) {
-    consola.warn('`nuxt export` has been deprecatrd! Please use `nuxt generate`.')
+    consola.warn('`nuxt export` has been deprecated! Please use `nuxt generate`.')
     return generate.run.call(this, ...args)
   }
 }
