+++ packages/cli/src/commands/start.js
@@ -1,6 +1,8 @@
 import { TARGETS } from '@nuxt/utils'
+import consola from 'consola'
 import { common, server } from '../options'
 import { showBanner } from '../utils/banner'
+import { serve } from '../utils/serve'
 
 export default {
   name: 'start',
@@ -12,9 +14,12 @@ export default {
   },
   async run (cmd) {
     const config = await cmd.getNuxtConfig({ dev: false, _start: true })
+
     if (config.target === TARGETS.static) {
-      throw new Error('You cannot use `nuxt start` with ' + TARGETS.static + ' target, please use `nuxt export` and `nuxt serve`')
+      consola.info('Serving static dist')
+      return serve(cmd)
     }
+
     const nuxt = await cmd.getNuxt(config)
 
     // Listen and show ready banner
